adding registration process; api account registration queue

+++ server/models/auth.js
@@ -28,7 +28,7 @@ var cryptPassword                        = function (password, callback) {
  * @param callback
  */
 var comparePassword                      = function (password, userPassword, callback) {
-    console.log("comparePassword init")
+    console.log("comparePassword init");
     bcrypt.compare(userPassword, password, function (err, isPasswordMatch) {
         console.log("compare error: ", err);
         console.log("compare isPasswordMatch: ", isPasswordMatch);
@@ -43,43 +43,52 @@ var comparePassword                      = function (password, userPassword, cal
  * @param callback
  */
 exports.check_usernamePassword           = function (obj, callback) {
-    db.query("select * from accounts where username=?", [obj.username], function (err, results) {
-        console.log("query error: ", err);
-        //console.log(results);
-        if (!err) {
-            if (results.length == 1) {
-                var record = results[0];
-                console.log(record);
-                if (record.status == 1) {
-                    // saved hash, entered password raw, callback
-                    comparePassword(record.password, obj.password, function (err, isPasswordMatch) {
-                        if (isPasswordMatch) {
-                            callback(
-                                {
-                                    status: 1,
-                                    meta  : {
-                                        account_guid: record.guid
-                                    }
+    db.getConnection(function (connErr, conn) {
+        if (!connErr) {
+            conn.query("select * from accounts where username=?", [obj.username], function (err, results) {
+                conn.release();
+                console.log("query error: ", err);
+                //console.log(results);
+                if (!err) {
+                    if (results.length == 1) {
+                        var record = results[0];
+                        console.log(record);
+                        if (record.status == 1) {
+                            // saved hash, entered password raw, callback
+                            comparePassword(record.password, obj.password, function (err, isPasswordMatch) {
+                                if (isPasswordMatch) {
+                                    callback(
+                                        {
+                                            status: 1,
+                                            meta  : {
+                                                account_guid: record.guid
+                                            }
+                                        }
+                                    );
                                 }
-                            );
+                                // invalid username or password (no user found)
+                                else callback({status: 0, code: 4001})
+                            })
                         }
-                        // invalid username or password (no user found)
-                        else callback({status: 0, code: 4001})
-                    })
+                        // account is not enabled
+                        else callback({status: 0, code: 4002});
+                    }
+                    else if (results.length > 1) {
+                        // error occurred where we have too many of the same username
+                        callback({status: 0, code: 5001})
+                    }
+                    // invalid username or password (no user found)
+                    else callback({status: 0, code: 4001})
                 }
-                // account is not enabled
-                else callback({status: 0, code: 4002});
-            }
-            else if (results.length > 1) {
-                // error occurred where we have too many of the same username
-                callback({status: 0, code: 5001})
-            }
-            // invalid username or password (no user found)
-            else callback({status: 0, code: 4001})
+                // db error occurred
+                else callback({status: 0, code: 5000, error: err});
+            });
         }
-        // db error occurred
-        else callback({status: 0, code: 5000, error: err});
-    });
+        else {
+            // db error occurred
+            callback({status: 0, code: 5000, error: connErr});
+        }
+    })
 };
 /**
  * EXPORTS METHOD: logout
@@ -88,10 +97,15 @@ exports.check_usernamePassword           = function (obj, callback) {
  * @param callback
  */
 exports.logout                           = function (opts, callback) {
-    db.query("delete from sessions where sessionId=?", [opts.sessionId], function (err) {
-        if (!err) callback({status: 1});
-        else callback({status: 0, error: err});
-    })
+    db.getConnection(function (connErr, conn) {
+        if (!connErr) {
+            conn.query("delete from sessions where sessionId=?", [opts.sessionId], function (err) {
+                conn.release();
+                if (!err) callback({status: 1});
+                else callback({status: 0, error: err});
+            })
+        }
+    });
 };
 /**
  * EXPORTS METHOD: makeTempSession
@@ -105,17 +119,22 @@ exports.makeTempSession                  = function (opts, callback) {
     var session_guid = uuid.v4();
     // make a crypted hash of the session guid to store as the temp login session
     cryptPassword(session_guid, function (err, sessionHash) {
-        db.query(
-            "insert into sessions set guid=?, account_guid=?, sessionId=?, dateCreated=now(), ipaddress=?, sessionType='temp'",
-            [session_guid, opts.account_guid, sessionHash, opts.remoteAddress],
-            function (err) {
-                if (!err) {
-                    callback({status: 1, sessionHash: sessionHash});
-                }
-                // db error occurred
-                else callback({status: 0, code: 5000, error: err});
+        db.getConnection(function (connErr, conn) {
+            if (!connErr) {
+                conn.query(
+                    "insert into sessions set guid=?, account_guid=?, sessionId=?, dateCreated=now(), ipaddress=?, sessionType='temp'",
+                    [session_guid, opts.account_guid, sessionHash, opts.remoteAddress],
+                    function (err) {
+                        conn.release();
+                        if (!err) {
+                            callback({status: 1, sessionHash: sessionHash});
+                        }
+                        // db error occurred
+                        else callback({status: 0, code: 5000, error: err});
+                    }
+                )
             }
-        )
+        });
     });
 };
 /**
@@ -130,17 +149,22 @@ exports.makeSession                      = function (opts, callback) {
     var session_guid = uuid.v4();
     // make a crypted hash of the session guid to store as the temp login session
     cryptPassword(session_guid, function (err, sessionHash) {
-        db.query(
-            "insert into sessions set guid=?, account_guid=?, org_guid=?, sessionId=?, dateCreated=now(), ipaddress=?",
-            [session_guid, opts.account_guid, opts.org_guid, sessionHash, opts.remoteAddress],
-            function (err) {
-                if (!err) {
-                    callback({status: 1, sessionHash: sessionHash});
-                }
-                // db error occurred
-                else callback({status: 0, code: 5000, error: err});
+        db.getConnection(function (connErr, conn) {
+            if (!connErr) {
+                conn.query(
+                    "insert into sessions set guid=?, account_guid=?, org_guid=?, sessionId=?, dateCreated=now(), ipaddress=?",
+                    [session_guid, opts.account_guid, opts.org_guid, sessionHash, opts.remoteAddress],
+                    function (err) {
+                        conn.release();
+                        if (!err) {
+                            callback({status: 1, sessionHash: sessionHash});
+                        }
+                        // db error occurred
+                        else callback({status: 0, code: 5000, error: err});
+                    }
+                )
             }
-        )
+        });
     });
 };
 /**
@@ -151,26 +175,31 @@ exports.makeSession                      = function (opts, callback) {
  * @param callback
  */
 exports.clearSession                     = function (opts, callback) {
-    var sql, inserts;
-    console.log("CLEAR SESSION: ", opts)
-    if (opts.sessionId) {
-        sql     = "delete from sessions where sessionId=?";
-        inserts = [opts.sessionId]
-    }
-    else {
-        sql     = "delete from sessions where account_guid=? and ipaddress=?";
-        inserts = [opts.account_guid, opts.remoteAddress]
-    }
-    sql = db.format(sql, inserts);
-    console.log("CLEAR SESSION: ", sql);
-    db.query(
-        sql,
-        function (err) {
-            if (!err) callback({status: 1});
-            // db error occurred
-            else callback({status: 0, code: 5000, error: err});
+    db.getConnection(function (connErr, conn) {
+        if (!connErr) {
+            var sql, inserts;
+            console.log("CLEAR SESSION: ", opts);
+            if (opts.sessionId) {
+                sql     = "delete from sessions where sessionId=?";
+                inserts = [opts.sessionId]
+            }
+            else {
+                sql     = "delete from sessions where account_guid=? and ipaddress=?";
+                inserts = [opts.account_guid, opts.remoteAddress]
+            }
+            sql = conn.format(sql, inserts);
+            console.log("CLEAR SESSION: ", sql);
+            conn.query(
+                sql,
+                function (err) {
+                    conn.release();
+                    if (!err) callback({status: 1});
+                    // db error occurred
+                    else callback({status: 0, code: 5000, error: err});
+                }
+            )
         }
-    )
+    });
 };
 /**
  * EXPORTS METHOD: checkValidSessionId
@@ -180,31 +209,36 @@ exports.clearSession                     = function (opts, callback) {
  * @param callback
  */
 exports.checkValidSessionId              = function (opts, callback) {
-    console.log("checkValidSessionId -> ", opts);
-    var sessionType = "";
-    if (opts.sessionType) sessionType = opts.sessionType;
-    var sql     = "select guid,sessionType from sessions where sessionId=? and sessionType=?",
-        inserts = [opts.sessionId, sessionType];
-    sql = db.format(sql, inserts);
-    console.log("checkValidSessionId SQL -> ", sql);
-    db.query(
-        sql,
-        function (err, results) {
-            if (!err)                                                                                               {
-                if (results.length < 2) {
-                    callback({status: 1});
+    db.getConnection(function (connErr, conn) {
+        if (!connErr) {
+            console.log("checkValidSessionId -> ", opts);
+            var sessionType = "";
+            if (opts.sessionType) sessionType = opts.sessionType;
+            var sql     = "select guid,sessionType from sessions where sessionId=? and sessionType=?",
+                inserts = [opts.sessionId, sessionType];
+            sql         = conn.format(sql, inserts);
+            console.log("checkValidSessionId SQL -> ", sql);
+            conn.query(
+                sql,
+                function (err, results) {
+                    conn.release();
+                    if (!err) {
+                        if (results.length < 2) {
+                            callback({status: 1});
+                        }
+                        else if (results.length > 1) {
+                            // error occurred where we have too many of the same username
+                            callback({status: 0, code: 5001, message: "Too many sessions returned"})
+                        }
+                        // invalid sessionId, none found
+                        //else callback({status: 0, code: 4003});
+                    }
+                    // db error occurred
+                    else callback({status: 0, code: 5000, error: err});
                 }
-                else if (results.length > 1) {
-                    // error occurred where we have too many of the same username
-                    callback({status: 0, code: 5001, message: "Too many sessions returned"})
-                }
-                // invalid sessionId, none found
-                //else callback({status: 0, code: 4003});
-            }
-            // db error occurred
-            else callback({status: 0, code: 5000, error: err});
+            )
         }
-    )
+    });
 };
 /**
  * EXPORTS METHOD: getCurrentSessionRecordBySession
@@ -214,22 +248,65 @@ exports.checkValidSessionId              = function (opts, callback) {
  * @param callback
  */
 exports.getCurrentSessionRecordBySession = function (opts, callback) {
-    db.query(
-        "select * from sessions where sessionId=?",
-        [opts.sessionId],
-        function (err, results) {
-            if (!err) {
-                if (results.length == 1) {
-                    callback({status: 1, session: results[0]});
+    db.getConnection(function (connErr, conn) {
+        if (!connErr) {
+            conn.query(
+                "select * from sessions where sessionId=?",
+                [opts.sessionId],
+                function (err, results) {
+                    conn.release();
+                    if (!err) {
+                        if (results.length == 1) {
+                            callback({status: 1, session: results[0]});
+                        }
+                        else if (results.length > 1) {
+                            // error occurred where we have too many of the same username
+                            callback({status: 0, code: 5001, message: "Too many sessions returned"})
+                        }
+                        else callback({status: 0, code: 4003});
+                    }
+                    // db error occurred
+                    else callback({status: 0, code: 5000, error: err});
                 }
-                else if (results.length > 1) {
-                    // error occurred where we have too many of the same username
-                    callback({status: 0, code: 5001, message: "Too many sessions returned"})
-                }
-                else callback({status: 0, code: 4003});
-            }
-            // db error occurred
-            else callback({status: 0, code: 5000, error: err});
+            );
         }
-    );
-}
+    });
+};
+exports.registerOrg                      = function (opts, callback) {
+};
+exports.createApiAccountRecord           = function (opts, callback) {
+    db.getConnection(function (connErr, conn) {
+        if (!connErr) {
+            var new_guid = uuid.v4(),
+                inserts  = [
+                    new_guid,
+                    opts.siteTitle,
+                    opts.siteUrl,
+                    opts.contactEmail,
+                    opts.address1,
+                    opts.address2,
+                    opts.city,
+                    opts.state,
+                    opts.postalCode,
+                    opts.firstName,
+                    opts.lastName,
+                    opts.remoteAddress
+                ];
+            conn.query(
+                "insert into api_accounts set guid=?,siteTitle=?,siteUrl=?,contactEmail=?,address1=?,address2=?,city=?,state=?,postalCode=?,firstName=?,lastName=?,remoteAddress=?,status='0'",
+                inserts,
+                function (err) {
+                    conn.release();
+                    if (!err) {
+                        callback({status: 1});
+                    }
+                    // db error occurred
+                    else callback({status: 0, code: 5000, error: err});
+                }
+            )
+            ;
+        }
+        // db error occurred
+        else callback({status: 0, code: 5000, error: connErr});
+    });
+};
